requirements.txt
Django
python-dotenv
psycopg2
django-tinymce
